refactor: unify code style
